+++ server/modules/serviceName.js
@@ -23,7 +23,7 @@ function format(environment, service, slice) {
   return `${environment}-${service}${formatSlice(slice)}`;
 }
 
-function format2(service, slice) {
+function formatSQN(service, slice) {
   validateRequired(/^[^-]+$/, 'service', service);
   validateOptional(/^[^-]+$/, 'slice', slice);
   return `${service}${formatSlice(slice)}`;
@@ -41,6 +41,6 @@ function parse(environmentQualifiedName) {
 
 module.exports = {
   format,
-  format2,
+  formatSQN,
   parse
 };