+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- APIs for synchronising service versions across environments
+    - POST /environment/{environment}/sync-services/jobs
+    - GET /environment/{environment}/sync-services/job/{JobId}
+
 ## [6.9.5] 2017-10-02
 
 ### Changed